upldate setting
