+++ .settings/org.eclipse.wst.common.component
@@ -1,9 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+        
     <wb-module deploy-name="JavaSpringFrameworkProject-0.0.1-SNAPSHOT">
+                
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
+                
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
+                
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-        <property name="context-root" value="JavaSpringFrameworkProject"/>
+                
         <property name="java-output-path" value="/JavaSpringFrameworkProject/build/classes"/>
+        <property name="context-root" value="LearnRun"/>
+            
     </wb-module>
+    
 </project-modules>
