+++ src/main/java/com/semi/learn/dao/Impl/MemberDaoImpl.java
@@ -1,0 +1,45 @@
+package com.semi.learn.dao.Impl;
+
+import org.apache.ibatis.session.SqlSession;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+import com.semi.learn.dao.MemberDao;
+
+@Repository
+public class MemberDaoImpl implements MemberDao{
+	// MyBatis 접근(생성)
+	@Autowired // 자동생성
+	SqlSession session;
+	
+	String ns = "Member."; // ns: namespace
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
