+++ src/main/java/com/semi/learn/dao/MemberDao.java
@@ -1,0 +1,5 @@
+package com.semi.learn.dao;
+
+public interface MemberDao {
+	
+}
