+++ src/main/java/com/semi/learn/dto/MemberDto.java
@@ -1,0 +1,7 @@
+package com.semi.learn.dto;
+
+import java.io.Serializable;
+
+public class MemberDto implements Serializable{
+	
+}
