+++ src/main/java/com/semi/learn/service/MemberService.java
@@ -1,0 +1,5 @@
+package com.semi.learn.service;
+
+public interface MemberService {
+	
+}
