--- src/main/java/mul/cam/a/controller/BbsController.java
@@ -1,39 +1,0 @@
-package mul.cam.a.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-
-import mul.cam.a.service.BbsService;
-
-@Controller
-public class BbsController {
-	@Autowired
-	BbsService service;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
