--- src/main/java/mul/cam/a/dao/Impl/MemberDaoImpl.java
@@ -1,45 +1,0 @@
-package mul.cam.a.dao.Impl;
-
-import org.apache.ibatis.session.SqlSession;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
-
-import mul.cam.a.dao.MemberDao;
-
-@Repository
-public class MemberDaoImpl implements MemberDao{
-	// MyBatis 접근(생성)
-	@Autowired // 자동생성
-	SqlSession session;
-	
-	String ns = "Member."; // ns: namespace
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
