--- src/main/java/mul/cam/a/dao/MemberDao.java
@@ -1,5 +1,0 @@
-package mul.cam.a.dao;
-
-public interface MemberDao {
-	
-}
