--- src/main/java/mul/cam/a/dto/BbsDto.java
@@ -1,7 +1,0 @@
-package mul.cam.a.dto;
-
-import java.io.Serializable;
-
-public class BbsDto implements Serializable{
-	
-}
