--- src/main/java/mul/cam/a/dto/MemberDto.java
@@ -1,7 +1,0 @@
-package mul.cam.a.dto;
-
-import java.io.Serializable;
-
-public class MemberDto implements Serializable{
-	
-}
