--- src/main/java/mul/cam/a/service/BbsService.java
@@ -1,30 +1,0 @@
-package mul.cam.a.service;
-
-public interface BbsService {
-	
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
