--- src/main/java/mul/cam/a/service/Impl/MemberServiceImpl.java
@@ -1,43 +1,0 @@
-package mul.cam.a.service.Impl;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import mul.cam.a.dao.MemberDao;
-import mul.cam.a.service.MemberService;
-
-@Service
-public class MemberServiceImpl implements MemberService {
-	// Dao 접근(생성)
-	@Autowired
-	MemberDao dao;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
