--- src/main/java/mul/cam/a/service/MemberService.java
@@ -1,5 +1,0 @@
-package mul.cam.a.service;
-
-public interface MemberService {
-	
-}
