+++ src/main/webapp/WEB-INF/spring/aop-context.xml
@@ -8,5 +8,5 @@
 	<aop:aspectj-autoproxy />
 	
 	<!-- aop 해당 클래스 생성 -->
-	<bean id="myAop" class="mul.cam.a.aop.AopProc" />
+	<bean id="myAop" class="com.semi.learn.aop.AopProc" />
 </beans>
