+++ src/main/webapp/WEB-INF/spring/servlet-context.xml
@@ -16,7 +16,7 @@
 	<bean id="titlesConfigurer" class="org.springframework.web.servlet.view.tiles3.TilesConfigurer">
 		<property name="definitions">
 			<list>
-				<value>/WEB-INF/views/layout/layouts.xml</value>
+				<value>/WEB-INF/views/layout/tiles.xml</value>
 			</list>
 		</property>
 	</bean>
@@ -33,7 +33,7 @@
 	</bean>
 	
 	<!-- 공통 패키지(경로) -->
-	<context:component-scan base-package="mul.cam.a"></context:component-scan>
+	<context:component-scan base-package="com.semi.learn"></context:component-scan>
 	
 	<!-- Ajax 사용허가 -->
 	<mvc:annotation-driven />
