+++ src/main/webapp/WEB-INF/views/layout/menu.jsp
@@ -1,10 +1,9 @@
 <%@page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 
-<h1>header</h1>
-<a href="/LearnRun/mainpage">메인으로 이동</a>
-&nbsp;
+<h1>menu</h1>
+
 <a href="/LearnRun/login">로그인으로 이동</a>
-&nbsp;
+<br>
 <a href="/LearnRun/clslist">게시판으로 이동</a>
-&nbsp;
+<br>
 <a href="/LearnRun/mypage/myCls">마이페이지로 이동</a>
