+++ src/main/webapp/WEB-INF/views/login/login.jsp
@@ -1,0 +1,3 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<h1>로그인</h1>
