--- src/main/webapp/WEB-INF/views/mainpage/menu.jsp
@@ -1,9 +1,0 @@
-<%@page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-
-<h1>menu</h1>
-
-<a href="login.do">로그인으로 이동</a>
-<br>
-<a href="bbslist.do">게시판으로 이동</a>
-<br>
-<a href="bbswrite.do">글쓰기로 이동</a>
