+++ src/main/webapp/WEB-INF/views/mypage/menu.jsp
@@ -1,0 +1,9 @@
+<%@page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+
+<h1>menu</h1>
+
+<a href="login">로그인으로 이동</a>
+<br>
+<a href="clslist">게시판으로 이동</a>
+<br>
+<a href="mypage/myCls">마이페이지로 이동</a>
