+++ src/main/webapp/WEB-INF/views/mypage/myCls.jsp
@@ -1,0 +1,4 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<h1>myCls</h1>
+<div>수강중인 강의</div>
