+++ src/main/webapp/WEB-INF/web.xml
@@ -34,7 +34,14 @@ version="4.0">
   
   <servlet-mapping>
 	<servlet-name>dispatcherServlet</servlet-name>
-	<url-pattern>*.do</url-pattern>
+	<url-pattern>/</url-pattern>
+  </servlet-mapping>
+  
+  <!-- The mappings for the JSP servlet -->
+  <servlet-mapping>
+    <servlet-name>jsp</servlet-name>
+    <url-pattern>*.jsp</url-pattern>
+    <url-pattern>*.jspx</url-pattern>
   </servlet-mapping>
   
   <context-param>
