+++ src/main/webapp/index.jsp
@@ -1,5 +1,5 @@
 <%@page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 
 <%
-	response.sendRedirect("mainpage.do");
+	response.sendRedirect("mainpage");
 %>
